Split out Azure files
