+++ src/Simulation/EntryPointDriver/Azure/Azure.cs
@@ -7,14 +7,10 @@ using static Microsoft.Quantum.EntryPointDriver.Driver;
 using Microsoft.Quantum.EntryPointDriver.Mock;
 using Microsoft.Quantum.Runtime;
 using Microsoft.Quantum.Simulation.Common.Exceptions;
-using Microsoft.Quantum.Simulation.Core;
 using System;
-using System.Collections.Immutable;
-using System.IO;
-using System.Linq;
 using System.Threading.Tasks;
 
-namespace Microsoft.Quantum.EntryPointDriver
+namespace Microsoft.Quantum.EntryPointDriver.Azure
 {
     using Environment = System.Environment;
 
@@ -233,195 +229,5 @@ namespace Microsoft.Quantum.EntryPointDriver
 
             public int Shots { get; set; }
         }
-    }
-
-    /// <summary>
-    /// A Q# entry point submission.
-    /// </summary>
-    /// <typeparam name="TIn">The entry point argument type.</typeparam>
-    /// <typeparam name="TOut">The entry point return type.</typeparam>
-    public sealed class QSharpSubmission<TIn, TOut>
-    {
-        /// <summary>
-        /// The entry point info.
-        /// </summary>
-        internal EntryPointInfo<TIn, TOut> EntryPointInfo { get; }
-
-        /// <summary>
-        /// The entry point argument.
-        /// </summary>
-        internal TIn Argument { get; }
-
-        /// <summary>
-        /// Creates a Q# entry point submission.
-        /// </summary>
-        /// <param name="entryPointInfo">The entry point info.</param>
-        /// <param name="argument">The entry point argument.</param>
-        public QSharpSubmission(EntryPointInfo<TIn, TOut> entryPointInfo, TIn argument) =>
-            (this.EntryPointInfo, this.Argument) = (entryPointInfo, argument);
-    }
-
-    /// <summary>
-    /// A QIR entry point submission.
-    /// </summary>
-    public sealed class QirSubmission
-    {
-        /// <summary>
-        /// The QIR bitcode stream.
-        /// </summary>
-        internal Stream QirStream { get; }
-
-        /// <summary>
-        /// The entry point name.
-        /// </summary>
-        internal string EntryPointName { get; }
-
-        /// <summary>
-        /// The entry point arguments.
-        /// </summary>
-        internal ImmutableList<Argument> Arguments { get; }
-
-        /// <summary>
-        /// Creates a QIR entry point submission.
-        /// </summary>
-        /// <param name="qirStream">The QIR bitcode stream.</param>
-        /// <param name="entryPointName">The entry point name.</param>
-        /// <param name="arguments">The entry point arguments.</param>
-        public QirSubmission(Stream qirStream, string entryPointName, ImmutableList<Argument> arguments) =>
-            (this.QirStream, this.EntryPointName, this.Arguments) = (qirStream, entryPointName, arguments);
-    }
-
-    /// <summary>
-    /// The information to show in the output after the job is submitted.
-    /// </summary>
-    public enum OutputFormat
-    {
-        /// <summary>
-        /// Show a friendly message with a URI that can be used to see the job results.
-        /// </summary>
-        FriendlyUri,
-
-        /// <summary>
-        /// Show only the job ID.
-        /// </summary>
-        Id
-    }
-
-    /// <summary>
-    /// Settings for a submission to Azure Quantum.
-    /// </summary>
-    public sealed class AzureSettings
-    {
-        /// <summary>
-        /// The subscription ID.
-        /// </summary>
-        public string? Subscription { get; set; }
-
-        /// <summary>
-        /// The resource group name.
-        /// </summary>
-        public string? ResourceGroup { get; set; }
-
-        /// <summary>
-        /// The workspace name.
-        /// </summary>
-        public string? Workspace { get; set; }
-
-        /// <summary>
-        /// The target device ID.
-        /// </summary>
-        public string? Target { get; set; }
-
-        /// <summary>
-        /// The storage account connection string.
-        /// </summary>
-        public string? Storage { get; set; }
-
-        /// <summary>
-        /// The Azure Active Directory authentication token.
-        /// </summary>
-        public string? AadToken { get; set; }
-
-        /// <summary>
-        /// The base URI of the Azure Quantum endpoint.
-        /// If both <see cref="BaseUri"/> and <see cref="Location"/> properties are not null, <see cref="BaseUri"/> takes precedence.
-        /// </summary>
-        public Uri? BaseUri { get; set; }
-
-        /// <summary>
-        /// The location to use with the default Azure Quantum endpoint.
-        /// If both <see cref="BaseUri"/> and <see cref="Location"/> properties are not null, <see cref="BaseUri"/> takes precedence.
-        /// </summary>
-        public string? Location { get; set; }
-
-        /// <summary>
-        /// The name of the submitted job.
-        /// </summary>
-        public string? JobName { get; set; }
-
-        /// <summary>
-        /// The number of times the program is executed on the target machine.
-        /// </summary>
-        public int Shots { get; set; }
-
-        /// <summary>
-        /// The information to show in the output after the job is submitted.
-        /// </summary>
-        public OutputFormat Output { get; set; }
-
-        /// <summary>
-        /// Validate the program and options, but do not submit to Azure Quantum.
-        /// </summary>
-        public bool DryRun { get; set; }
-
-        /// <summary>
-        /// Show additional information about the submission.
-        /// </summary>
-        public bool Verbose { get; set; }
-
-        /// <summary>
-        /// Creates a <see cref="Workspace"/> based on the settings.
-        /// </summary>
-        /// <returns>The <see cref="Workspace"/> based on the settings.</returns>
-        internal Workspace CreateWorkspace()
-        {
-            if (BaseUri != null)
-            {
-                return AadToken is null
-                    ? new Workspace(Subscription, ResourceGroup, Workspace, baseUri: BaseUri)
-                    : new Workspace(Subscription, ResourceGroup, Workspace, AadToken, baseUri: BaseUri);
-            }
-            else if (Location != null)
-            {
-                return AadToken is null
-                    ? new Workspace(Subscription, ResourceGroup, Workspace, location: NormalizeLocation(Location))
-                    : new Workspace(Subscription, ResourceGroup, Workspace, AadToken, location: NormalizeLocation(Location));
-            }
-            else
-            {
-                return AadToken is null
-                    ? new Workspace(Subscription, ResourceGroup, Workspace, baseUri: null)
-                    : new Workspace(Subscription, ResourceGroup, Workspace, AadToken, baseUri: null);
-            }
-        }
-
-        public override string ToString() => string.Join(
-            Environment.NewLine,
-            $"Subscription: {Subscription}",
-            $"Resource Group: {ResourceGroup}",
-            $"Workspace: {Workspace}",
-            $"Target: {Target}",
-            $"Storage: {Storage}",
-            $"AAD Token: {AadToken}",
-            $"Base URI: {BaseUri}",
-            $"Location: {Location}",
-            $"Job Name: {JobName}",
-            $"Shots: {Shots}",
-            $"Output: {Output}",
-            $"Dry Run: {DryRun}",
-            $"Verbose: {Verbose}");
-
-        internal static string NormalizeLocation(string location) =>
-            string.Concat(location.Where(c => !char.IsWhiteSpace(c))).ToLower();
     }
 }
