+++ src/Simulation/EntryPointDriver/Azure/OutputFormat.cs
@@ -1,0 +1,21 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace Microsoft.Quantum.EntryPointDriver.Azure
+{
+    /// <summary>
+    /// The information to show in the output after the job is submitted.
+    /// </summary>
+    public enum OutputFormat
+    {
+        /// <summary>
+        /// Show a friendly message with a URI that can be used to see the job results.
+        /// </summary>
+        FriendlyUri,
+
+        /// <summary>
+        /// Show only the job ID.
+        /// </summary>
+        Id
+    }
+}
