+++ src/Simulation/EntryPointDriver/Driver.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using Microsoft.Quantum.EntryPointDriver.Azure;
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
